Unrelated fix (input stuck under container)
