+++ packages/twenty-front/src/modules/object-record/record-table/record-table-cell/components/RecordTableCellContainer.tsx
@@ -6,10 +6,14 @@ import { useGetButtonIcon } from '@/object-record/record-field/hooks/useGetButto
 import { useIsFieldEmpty } from '@/object-record/record-field/hooks/useIsFieldEmpty';
 import { useIsFieldInputOnly } from '@/object-record/record-field/hooks/useIsFieldInputOnly';
 import { RecordTableCellContext } from '@/object-record/record-table/contexts/RecordTableCellContext';
+import { RecordTableRowContext } from '@/object-record/record-table/contexts/RecordTableRowContext';
 import { useGetIsSomeCellInEditModeState } from '@/object-record/record-table/hooks/internal/useGetIsSomeCellInEditMode';
 import { useOpenRecordTableCell } from '@/object-record/record-table/record-table-cell/hooks/useOpenRecordTableCell';
+import { useSetCurrentRowSelected } from '@/object-record/record-table/record-table-row/hooks/useSetCurrentRowSelected';
 import { isSoftFocusUsingMouseState } from '@/object-record/record-table/states/isSoftFocusUsingMouseState';
 import { IconArrowUpRight } from '@/ui/display/icon';
+import { contextMenuIsOpenState } from '@/ui/navigation/context-menu/states/contextMenuIsOpenState';
+import { contextMenuPositionState } from '@/ui/navigation/context-menu/states/contextMenuPositionState';
 import { HotkeyScope } from '@/ui/utilities/hotkey/types/HotkeyScope';
 
 import { CellHotkeyScopeContext } from '../../contexts/CellHotkeyScopeContext';
@@ -23,6 +27,12 @@ import { RecordTableCellButton } from './RecordTableCellButton';
 import { RecordTableCellDisplayMode } from './RecordTableCellDisplayMode';
 import { RecordTableCellEditMode } from './RecordTableCellEditMode';
 import { RecordTableCellSoftFocusMode } from './RecordTableCellSoftFocusMode';
+
+const StyledTd = styled.td<{ isSelected: boolean; isInEditMode: boolean }>`
+  background: ${({ isSelected, theme }) =>
+    isSelected ? theme.accent.quaternary : theme.background.primary};
+  z-index: ${({ isInEditMode }) => (isInEditMode ? '4 !important' : '3')};
+`;
 
 const StyledCellBaseContainer = styled.div`
   align-items: center;
@@ -80,6 +90,23 @@ export const RecordTableCellContainer = ({
     openTableCell();
   };
 
+  const { isSelected } = useContext(RecordTableRowContext);
+
+  const setContextMenuPosition = useSetRecoilState(contextMenuPositionState);
+  const setContextMenuOpenState = useSetRecoilState(contextMenuIsOpenState);
+
+  const { setCurrentRowSelected } = useSetCurrentRowSelected();
+
+  const handleContextMenu = (event: React.MouseEvent) => {
+    event.preventDefault();
+    setCurrentRowSelected(true);
+    setContextMenuPosition({
+      x: event.clientX,
+      y: event.clientY,
+    });
+    setContextMenuOpenState(true);
+  };
+
   const handleContainerMouseEnter = () => {
     if (!isHovered && !isSomeCellInEditMode) {
       setIsHovered(true);
@@ -109,46 +136,52 @@ export const RecordTableCellContainer = ({
     (!isFirstColumn || !isEmpty);
 
   return (
-    <CellHotkeyScopeContext.Provider
-      value={editHotkeyScope ?? DEFAULT_CELL_SCOPE}
+    <StyledTd
+      isSelected={isSelected}
+      onContextMenu={(event) => handleContextMenu(event)}
+      isInEditMode={isCurrentTableCellInEditMode}
     >
-      <StyledCellBaseContainer
-        onMouseEnter={handleContainerMouseEnter}
-        onMouseLeave={handleContainerMouseLeave}
+      <CellHotkeyScopeContext.Provider
+        value={editHotkeyScope ?? DEFAULT_CELL_SCOPE}
       >
-        {isCurrentTableCellInEditMode ? (
-          <RecordTableCellEditMode
-            editModeHorizontalAlign={editModeHorizontalAlign}
-            editModeVerticalPosition={editModeVerticalPosition}
-          >
-            {editModeContent}
-          </RecordTableCellEditMode>
-        ) : hasSoftFocus ? (
-          <>
-            {showButton && (
-              <RecordTableCellButton
-                onClick={handleButtonClick}
-                Icon={buttonIcon}
-              />
-            )}
-            <RecordTableCellSoftFocusMode>
-              {editModeContentOnly ? editModeContent : nonEditModeContent}
-            </RecordTableCellSoftFocusMode>
-          </>
-        ) : (
-          <>
-            {showButton && (
-              <RecordTableCellButton
-                onClick={handleButtonClick}
-                Icon={buttonIcon}
-              />
-            )}
-            <RecordTableCellDisplayMode>
-              {editModeContentOnly ? editModeContent : nonEditModeContent}
-            </RecordTableCellDisplayMode>
-          </>
-        )}
-      </StyledCellBaseContainer>
-    </CellHotkeyScopeContext.Provider>
+        <StyledCellBaseContainer
+          onMouseEnter={handleContainerMouseEnter}
+          onMouseLeave={handleContainerMouseLeave}
+        >
+          {isCurrentTableCellInEditMode ? (
+            <RecordTableCellEditMode
+              editModeHorizontalAlign={editModeHorizontalAlign}
+              editModeVerticalPosition={editModeVerticalPosition}
+            >
+              {editModeContent}
+            </RecordTableCellEditMode>
+          ) : hasSoftFocus ? (
+            <>
+              {showButton && (
+                <RecordTableCellButton
+                  onClick={handleButtonClick}
+                  Icon={buttonIcon}
+                />
+              )}
+              <RecordTableCellSoftFocusMode>
+                {editModeContentOnly ? editModeContent : nonEditModeContent}
+              </RecordTableCellSoftFocusMode>
+            </>
+          ) : (
+            <>
+              {showButton && (
+                <RecordTableCellButton
+                  onClick={handleButtonClick}
+                  Icon={buttonIcon}
+                />
+              )}
+              <RecordTableCellDisplayMode>
+                {editModeContentOnly ? editModeContent : nonEditModeContent}
+              </RecordTableCellDisplayMode>
+            </>
+          )}
+        </StyledCellBaseContainer>
+      </CellHotkeyScopeContext.Provider>
+    </StyledTd>
   );
 };
